Fixed and refactor
