+++ src/App.tsx
@@ -1,72 +1,29 @@
 /* eslint-disable jsx-a11y/label-has-associated-control */
 /* eslint-disable jsx-a11y/control-has-associated-label */
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import { UserWarning } from './UserWarning';
-import { deleteTodo, getTodos, USER_ID } from './api/todos';
-import { Todo } from './types/Todo';
 import { TodoList } from './components/TodoList';
 import { Footer } from './components/Footer';
 import { Header } from './components/Header';
 import { ErrorNotification } from './components/ErrorNotification';
-import { FilterStatus } from './types/FilterStatus';
+import { USER_ID } from './api/todos';
+import { useTodos } from './hook/useTodos';
 
 export const App: React.FC = () => {
-  const [todos, setTodos] = useState<Todo[]>([]);
-  const [filteredTodos, setFilteredTodos] = useState<Todo[]>([]);
-  const [filterStatus, setFilterStatus] = useState<FilterStatus>(
-    FilterStatus.ALL,
-  );
-  const [error, setError] = useState<string>('');
-  const [tempTodo, setTempTodo] = useState<Todo | null>(null);
-  const [deleteTodoId, setDeleteTodoId] = useState<number | null>(null);
-  const [isDeleted, setIsDeleted] = useState<boolean>(false);
-
-  useEffect(() => {
-    setError('');
-    getTodos()
-      .then(data => setTodos(data))
-      .catch(() => setError('Unable to load todos'));
-  }, []);
-
-  useEffect(() => {
-    let newFilterTodos = [...todos];
-
-    switch (filterStatus) {
-      case FilterStatus.ACTIVE:
-        newFilterTodos = todos.filter(todo => !todo.completed);
-        break;
-      case FilterStatus.COMPLETED:
-        newFilterTodos = todos.filter(todo => todo.completed);
-        break;
-    }
-
-    setFilteredTodos(newFilterTodos);
-  }, [filterStatus, todos]);
-
-  const removeTodo = (id: number) => {
-    setDeleteTodoId(id);
-    setIsDeleted(false);
-
-    return deleteTodo(id)
-      .then(() => {
-        setTodos(prevTodos => prevTodos.filter(prevTodo => prevTodo.id !== id));
-      })
-      .catch(() => setError('Unable to delete a todo'))
-      .finally(() => {
-        setDeleteTodoId(null);
-        setIsDeleted(true);
-      });
-  };
-
-  const deleteAllCompletedTodos = () => {
-    const completedTodos = todos.filter(todo => todo.completed);
-
-    if (completedTodos.length === 0) {
-      return;
-    }
-
-    Promise.allSettled(completedTodos.map(todo => removeTodo(todo.id)));
-  };
+  const {
+    todos,
+    setTodos,
+    filterStatus,
+    setFilterStatus,
+    error,
+    setError,
+    tempTodo,
+    setTempTodo,
+    deleteTodosId,
+    filteredTodos,
+    removeTodo,
+    deleteAllCompletedTodos,
+  } = useTodos();
 
   if (!USER_ID) {
     return <UserWarning />;
@@ -79,10 +36,10 @@ export const App: React.FC = () => {
       <div className="todoapp__content">
         <Header
           filteredTodos={filteredTodos}
+          error={error}
           setError={setError}
           setTodos={setTodos}
           setTempTodo={setTempTodo}
-          isDeleted={isDeleted}
         />
         {todos?.length > 0 && (
           <>
@@ -90,7 +47,7 @@ export const App: React.FC = () => {
               filteredTodos={filteredTodos}
               tempTodo={tempTodo}
               removeTodo={removeTodo}
-              deleteTodoId={deleteTodoId}
+              deleteTodosId={deleteTodosId}
             />
             <Footer
               setFilterStatus={setFilterStatus}
