+++ src/components/ErrorNotification/ErrorNotification.tsx
@@ -1,3 +1,4 @@
+import classNames from 'classnames';
 import { useEffect } from 'react';
 
 type Props = {
@@ -23,7 +24,10 @@ export const ErrorNotification: React.FC<Props> = ({ error, setError }) => {
   return (
     <div
       data-cy="ErrorNotification"
-      className={`notification is-danger is-light has-text-weight-normal ${error ? '' : 'hidden'}`}
+      className={classNames(
+        'notification is-danger is-light has-text-weight-normal',
+        { hidden: !error },
+      )}
     >
       <button data-cy="HideErrorButton" type="button" className="delete" />
       {error}
