+++ src/components/Header/Header.tsx
@@ -1,90 +1,40 @@
-import { useEffect, useRef, useState } from 'react';
 import { Todo } from '../../types/Todo';
-import { addTodo, USER_ID } from '../../api/todos';
+import classNames from 'classnames';
+import { useNewTodo } from '../../hook/useNewTodo';
 
 type Props = {
   filteredTodos: Todo[];
+  error: string;
   setError: (value: string) => void;
   setTodos: (fn: SetTodosFuncion) => void;
   setTempTodo: (tempTask: Todo | null) => void;
-  isDeleted: boolean;
 };
 
 type SetTodosFuncion = (todo: Todo[]) => Todo[];
 
 export const Header: React.FC<Props> = ({
   filteredTodos,
+  error,
   setError,
   setTodos,
   setTempTodo,
-  isDeleted,
 }) => {
-  const hasAllTodosCompleted = filteredTodos.every(todo => todo.completed);
-  const [title, setTitle] = useState<string>('');
-  const [isLoading, setIsLoading] = useState<boolean>(false);
-
-  // eslint-disable-next-line @typescript-eslint/naming-convention, @typescript-eslint/no-unused-vars
-  const inputRef = useRef<HTMLInputElement>(null);
-
-  useEffect(() => {
-    inputRef.current?.focus();
-  }, [isDeleted]);
-
-  const handleTitle = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setTitle(event.target.value);
-  };
-
-  const addNewTodo = (newTask: Omit<Todo, 'id'>) => {
-    const tempTask: Todo = {
-      id: 0,
-      userId: USER_ID,
-      title: newTask.title,
-      completed: false,
-    };
-
-    setTempTodo(tempTask);
-    setIsLoading(true);
-
-    addTodo(newTask)
-      .then(data => {
-        setTodos(prevTodos => [...prevTodos, data]);
-        setTempTodo(null);
-        setTitle('');
-      })
-      .catch(() => {
-        setError('Unable to add a todo');
-        setTempTodo(null);
-      })
-      .finally(() => {
-        setIsLoading(false);
-        setTimeout(() => inputRef.current?.focus(), 0);
-      });
-  };
-
-  const handleNewTodo = (event: React.FormEvent) => {
-    event.preventDefault();
-    const trimTitle = title.trim();
-
-    if (!trimTitle) {
-      setError('Title should not be empty');
-
-      return;
-    }
-
-    const newTodoTask = {
-      userId: USER_ID,
-      title: trimTitle,
-      completed: false,
-    };
-
-    addNewTodo(newTodoTask);
-  };
+  const {
+    title,
+    isLoading,
+    inputRef,
+    handleTitle,
+    handleNewTodo,
+    hasAllTodosCompleted,
+  } = useNewTodo({ filteredTodos, error, setError, setTodos, setTempTodo });
 
   return (
     <header className="todoapp__header">
       <button
         type="button"
-        className={`todoapp__toggle-all ${hasAllTodosCompleted ? 'active' : ''}`}
+        className={classNames('todoapp__toggle-all', {
+          active: hasAllTodosCompleted,
+        })}
         data-cy="ToggleAllButton"
       />
